refactor(tests): reorganize unit tests module  structure
