+++ tests/agents/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for strix.agents module."""
