+++ tests/interface/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for strix.interface module."""
