+++ tests/tools/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for strix.tools module."""
